Add SMB3.1 dialect security part1
For details on SMB3.1 see:
http://download.microsoft.com/download/C/6/C/C6C3C6F1-E84A-44EF-82A9-49BD3AAD8F58/Windows/%5BMS-SMB2-Diff%5D.pdf

This first part of the SMB3.1 dialect enablement adds mount
parsing check for the new protocol version (vers=3.1)
and updates negotiate protocol to handle the new negotiate
format and to send the supported hash and CMAC mechanisms.
Although it works for mount (when signing disabled) it does
not completely parse and save the encryption/hashing mechanisms
(which needs to be finished).

A followon patch will save the mechanisms for the
session and calculate the hashes on subsequent SMBs.
This will be necessary when signing is enabled.

This is sufficient for doing mount without signing enabled
though.

Signed-off-by: Steve French <smfrench@gmail.com>